Update iou to use get_overlapping_bounding_boxes
