+++ src/cell_tracking_metrics/matchers/compute_overlap.py
@@ -6,6 +6,38 @@ Licensed under The MIT License [see LICENSE for details]
 Copyright (c) 2015 Microsoft
 """
 import numpy as np
+from skimage.measure import regionprops
+
+
+def get_overlapping_bounding_boxes(gt_frame, res_frame):
+    gt_props = regionprops(gt_frame.astype(np.uint16))
+    gt_boxes = [np.array(gt_prop.bbox) for gt_prop in gt_props]
+    gt_boxes = np.array(gt_boxes).astype(np.float64)
+    gt_box_labels = np.asarray(
+        [int(gt_prop.label) for gt_prop in gt_props], dtype=np.uint16
+    )
+
+    res_props = regionprops(res_frame.astype(np.uint16))
+    res_boxes = [np.array(res_prop.bbox) for res_prop in res_props]
+    res_boxes = np.array(res_boxes).astype(np.float64)
+    res_box_labels = np.asarray(
+        [int(res_prop.label) for res_prop in res_props], dtype=np.uint16
+    )
+
+    if gt_frame.ndim == 3:
+        overlaps = compute_overlap_3D(gt_boxes, res_boxes)
+    else:
+        overlaps = compute_overlap(
+            gt_boxes, res_boxes
+        )  # has the form [gt_bbox, res_bbox]
+
+    # Find the bboxes that have overlap at all (ind_ corresponds to box number - starting at 0)
+    ind_gt, ind_res = np.nonzero(overlaps)
+    ind_gt = np.asarray(ind_gt, dtype=np.uint16)
+    ind_res = np.asarray(ind_res, dtype=np.uint16)
+    overlapping_gt_labels = gt_box_labels[ind_gt]
+    overlapping_res_labels = res_box_labels[ind_res]
+    return overlapping_gt_labels, overlapping_res_labels
 
 
 def compute_overlap(boxes: np.ndarray, query_boxes: np.ndarray) -> np.ndarray:
